Add AjoutRecetteScreen Ajout dans MainActivity.kt
Issue :

button collapse

+++ app/src/main/java/com/example/chief/data/model/Ingredient.kt
@@ -2,6 +2,6 @@ package com.example.chief.data.model
 
 data class Ingredient(
     val nom: String,
-    val quantite: Double,
+    val quantite: Int,
     val unite: String
 )
